+++ packages/kitchensink/service-fetchq/index.js
@@ -1,6 +1,8 @@
 const { runHookApp } = require('@forrestjs/hooks');
-const serviceFastify = require('@forrestjs/service-fastify');
+const serviceLogger = require('@forrestjs/service-logger');
 const serviceFetchq = require('@forrestjs/service-fetchq');
+const serviceFastify = require('@forrestjs/service-fastify');
+const serviceFastifyHealthz = require('@forrestjs/service-fastify-healthz');
 
 const featureHome = require('./feature-home');
 const featureQ1 = require('./feature-q1');
@@ -10,11 +12,14 @@ runHookApp({
   settings: {
     fetchq: {
       logLevel: 'error',
-      pool: {
-        max: 1,
-      },
+      pool: { max: 1 },
     },
   },
-  services: [serviceFetchq, serviceFastify],
+  services: [
+    serviceLogger,
+    serviceFetchq,
+    serviceFastify,
+    serviceFastifyHealthz,
+  ],
   features: [featureHome, featureQ1],
 }).catch((err) => console.error(err));